Generator adds therubyracer to the Gemfile
Previously we required users to manually add therubyracer to their
Gemfile. There was really no reason for this since we can just have
the generator do it for them. We have chosen not to make therubyracer
a gem dependency in the gemspec because it is MRI-specific and
therefore will not work with JRuby.